fix(ci): python-coverage-shards — split overlay test part_001 (1530 lines) under the 1500 first-party file-line guardrail
The reviewer-fix commits grew test_claude_auth_overlay_part_001.py to 1530
lines, tripping test_first_party_code_files_stay_under_line_limit. Move the
three legacy completeness-marker reap tests into a new part_006, importing the
shared FakeOverlayMounter / _recording_mknod / _seed_* helpers from part_001
(the established cross-part import pattern), bringing part_001 to 1387 lines.
